dynamic_document_update/requirements.txt
lxml==4.9.3
mysql-connector-python==8.1.0
protobuf==4.21.12
python-decouple==3.8
python-docx==1.0.1
schedule==1.2.1
typing_extensions==4.8.0
